% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovCoagBaseValidation.R
\docType{package}
\name{CovCoagBaseValidation}
\alias{CovCoagBaseValidation}
\title{CovCoagBaseValidation}
\description{
A package for validating prediction models using data in the OMOP CDM
}
